+++ app/controllers/sessions_controller.rb
@@ -7,6 +7,8 @@ class SessionsController < ApplicationController
   def create
     if registered_user
       create_session
+      assign_active_directory_user_id
+
       redirect_to root_path
     else
       redirect_to sign_in_path, notice: I18n.t("unknown_user.message", email_address: authenticated_user_email_address)
@@ -32,8 +34,18 @@ class SessionsController < ApplicationController
     session[:user_id] = registered_user.id
   end
 
+  def assign_active_directory_user_id
+    return if registered_user.active_directory_user_id.present?
+
+    registered_user.update(active_directory_user_id: active_directory_user_id)
+  end
+
   def authenticated_user_info
     request.env["omniauth.auth"].info
+  end
+
+  def active_directory_user_id
+    request.env["omniauth.auth"]["uid"]
   end
 
   def authenticated_user_email_address