+++ db/migrate/20220901125956_add_active_directory_user_id_to_user.rb
@@ -1,0 +1,5 @@
+class AddActiveDirectoryUserIdToUser < ActiveRecord::Migration[7.0]
+  def change
+    add_column :users, :active_directory_user_id, :string
+  end
+end