+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2022_08_30_145914) do
+ActiveRecord::Schema[7.0].define(version: 2022_09_01_125956) do
   create_table "actions", id: :uuid, default: -> { "newid()" }, force: :cascade do |t|
     t.string "title", null: false
     t.integer "order", null: false
@@ -94,6 +94,7 @@ ActiveRecord::Schema[7.0].define(version: 2022_08_30_145914) do
     t.boolean "regional_delivery_officer", default: false, null: false
     t.string "first_name"
     t.string "last_name"
+    t.string "active_directory_user_id"
     t.index ["email"], name: "index_users_on_email", unique: true
   end