+++ spec/support/sign_in_helpers.rb
@@ -3,7 +3,8 @@ module SignInHelpers
     OmniAuth.config.mock_auth[:azure_activedirectory_v2] = OmniAuth::AuthHash.new({
       info: {
         email: email_address
-      }
+      },
+      uid: "b5095c3e-0141-4478-81ad-99d0fbd087ed"
     })
     OmniAuth.config.test_mode = true
   end