+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Add a bulk user import Rake task.
+- Collect Azure Active Directory User ID on sign in.
 
 ### Changed